feat: Consistent imports rule

+++ packages/backend-core/src/util/base64url.ts
@@ -36,7 +36,7 @@ export function parse(input: string): Uint8Array {
   // Allocate the output:
   const out = new Uint8Array(
     ((end * base64UrlEncoding.bits) / 8) | 0
-  ) as Uint8Array;
+  ) ;
 
   // Parse the data:
   let bits = 0; // Number of bits currently in the buffer
